+++ lib/jin_widget_helper.dart
@@ -1,7 +1,14 @@
 library jin_widget_helper;
 
-export './stream_handler.dart';
-export './page_navigator.dart';
-export './small_icon_button.dart';
-export './widget_helper.dart';
-export './action_button.dart';
+export './src/stream_handler.dart';
+export './src/page_navigator.dart';
+export './src/small_icon_button.dart';
+export './src/widget_helper.dart';
+export './src/action_button.dart';
+export './src/badge_button.dart';
+export './src/color_utils.dart';
+export './src/custom_back_button.dart';
+export './src/form_validator.dart';
+export './src/app_extension.dart';
+export './src/connection_checker.dart';
+export './src/general_dialog.dart';