+++ lib/src/widget_helper.dart
@@ -1,0 +1,27 @@
+import 'dart:typed_data';
+import 'dart:ui';
+
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+class WidgetHelper {
+  ///Create a sizedBox with set height
+  static Widget verticalSpace([double height = 8]) {
+    return SizedBox(height: height);
+  }
+
+  ///Create a sizedBox with set width
+  static Widget horizontalSpace([double width = 8]) {
+    return SizedBox(width: width);
+  }
+
+  static Future<Uint8List> getBytesFromAsset(String path, int width) async {
+    ByteData data = await rootBundle.load(path);
+    Codec codec = await instantiateImageCodec(data.buffer.asUint8List(),
+        targetWidth: width);
+    FrameInfo fi = await codec.getNextFrame();
+    return (await fi.image.toByteData(format: ImageByteFormat.png))
+        .buffer
+        .asUint8List();
+  }
+}